requirements.txt
certifi==2018.8.24
chardet==3.0.4
idna==2.7
lxml==4.2.5
numpy==1.15.1
pandas==0.23.4
py4j==0.10.7
pyenchant==2.0.0
pyspark==2.3.1
python-dateutil==2.7.3
pytrends==4.4.0
pytz==2018.5
requests==2.19.1
six==1.11.0
urllib3==1.23
